pisi 2.4_alpha4
This is the forth alpha for the upcoming 2.4.x releases. This release
introduces two features as well as bug fixes and translation updates.

The first feature which is added in this release is an API addition for
use in buildfarm. With a new method, it is now possible to determine
the old packages that will be used to create delta packages from.

"Build flags" is another feature introduced in this release. Build flags
can be used to enable/disable some operations/packages at build time.
Currently, two flags are accepted: noDebug and noDelta.

Changes since 2.4_alpha3
------------------------
fatih (11):
      build: Reload actions script for each build type
      installdb: Parse directory names correctly
      l10n: Update messages
      build: Do not allow duplicate path values in spec file
      build: Fail if a path does not start with a slash
      build: Add a method to determine delta packages to be created
      l10n: Update messages
      index: Ignore debug packages of obsoleted packages
      build,specfile: Implement support for build flags
      dtd,rng: Accept BuildFlags tag inside Package tags
      l10n: Update Turkish translations

igungor (1):
      l10n: Revert the latest Italian translation which breaks pisi.

transifex (5):
      l10n: Updated Italian (it) translation to 81% (Hudaverdi Sarialtin)
      l10n: Updated French (fr) translation to 100% (Philippe Svetoslavsky)
      l10n: Updated Dutch (Flemish) (nl) translation to 100% (Antoon Tolboom)
      l10n: Updated Dutch (Flemish) (nl) translation to 100% (Antoon Tolboom)
      l10n: Updated French (fr) translation to 100% (Philippe Svetoslavsky)