pisi 2.4_beta1
This is the first beta for the upcoming 2.4.x releases. This release
includes enhancements, important bug fixes, and translation updates.

The most notable change is the parallelization of index module. This
reduces the time spent when indexing large repositories. Another
important change is the fix for failure when a directory in a package
becomes a symlink.

There are also behavioural changes introduced in this release:

 * delta: If all files in a package have changed, delta is not created
   unless requested explicitly.
 * build: Additional source files are now copied before applying
   patches.
 * cli: add-repo command now fails when a repo name/uri is not given.